+++ example.env
@@ -3,3 +3,4 @@ OWNER="YOUR_USER_ID"
 DEV_ERROR_CHANNEL_ID="YOUR_CHANNEL_ID"
 TOPGG_TOKEN="YOUR_TOPGG_TOKEN"
 ENABLE_MEDIA_FETCHING="false"
+TATSU_TOKEN="YOUR_TATSU_TOKEN"
